Basic Flask Structure and adapted header tabs
Implemented basic flask structure as per the week 7 lecture on Flask.

Also implemented "header.html" which is shared among the new tab pages, i.e. "home.html", "get.html", etc. This adapts the first version of the header tabs which was created without being implemented in a server context.

Clicking on a tab in the header now runs a server get request to fetch the route of the desired endpoint. Default is still the home page.