Header changes
Better formatting of header to better sync with steam style.

More uniform padding and spacing. Better text location in the header.

Colour change on hover, pointer on hover.

Not on Steam but implemented an underline animation on hover. Can take this out, just thought it was cool.

Currently commented out of .active-tab is the underline, kinda ugly but can always add back in.