SteamID can be linked in profile
database updated to link relationship between user and steam user via id key

Can link, edit and update steam id in profile page

by linking or updating, the GetPlayerSummaries(steam_id) function is run from steam_utils. This fills out the Steam_User table in db and links it to the USer table with the corresponding id (not to be mixed up with steam_id)

as GetPlayerSummaries(steam_id) is a api call, it takes a few seconds. Will implement a loading screen/wheel